Also add a shim-signed.postrm
Similarly to the postinst, we need to call grub-install to ensure the
boot chain is correct. If we're being remove, *again* work out the
right grub EFI target arch and the grub-install options so we can call
grub-install safely.

Also: grub-install won't clean up after us if we've been uninstalled,
so also go and remove the shim binary from the ESP.